requirements.txt
scapy
plyer
pytz
datetime
python-geoip-python3
python-geoip-geolite2